# Improve query performance of retrieve latest measures by area 